refactor: 도메인 모델 패키지 이동에 따른 임포트 경로 업데이트 및 코드 정돈
- 도메인 모델의 패키지 위치 변경(`domain.model` -> `domain.model.author`)에 따른 `Detail`, `ListItem`, `ProcessingMethod`, `SearchedSong` 등의 임포트 경로 수정
- `AfternoteEditorViewModel` 내 예외 처리 로직의 `when` 문 문법 개선 (Kotlin 1.9+ 스타일 적용)
- `AfternoteEditorScreen`에 `FlowPreview` 어노테이션 추가 및 불필요한 주석 제거
- `AfternoteEditorMapper` 내 미사용 주석 제거 및 가독성 개선을 위한 코드 정리
