+++ feature/afternote/presentation/src/main/kotlin/com/afternote/feature/afternote/presentation/author/editor/model/RegisterAfternotePayload.kt
@@ -1,9 +1,9 @@
 package com.afternote.feature.afternote.presentation.author.editor.model
-import com.afternote.feature.afternote.domain.model.ProcessingMethod
+import com.afternote.feature.afternote.domain.model.author.ProcessingMethod
 
 /**
  * Payload passed when the user taps "등록" on the afternote edit screen.
- * Used to build a new [com.afternote.feature.afternote.domain.model.ListItem] for create/update requests.
+ * Used to build a new [com.afternote.feature.afternote.domain.model.author.ListItem] for create/update requests.
  *
  * @param atmosphere Memorial(PLAYLIST) only: "남기고 싶은 당부" text sent as playlist.atmosphere in PATCH.
  */
