+++ feature/afternote/presentation/src/main/kotlin/com/afternote/feature/afternote/presentation/author/editor/playlist/SongMapper.kt
@@ -1,5 +1,5 @@
 package com.afternote.feature.afternote.presentation.author.editor.playlist
-import com.afternote.feature.afternote.domain.model.playlist.SearchedSong
+import com.afternote.feature.afternote.domain.model.author.playlist.SearchedSong
 import com.afternote.feature.afternote.presentation.shared.model.PlaylistSongDisplay
 
 fun SearchedSong.toDisplay() =
