+++ feature/afternote/presentation/src/main/kotlin/com/afternote/feature/afternote/presentation/author/home/model/AfternoteHomeUiState.kt
@@ -1,7 +1,7 @@
 package com.afternote.feature.afternote.presentation.author.home.model
 
 import com.afternote.core.ui.scaffold.bottombar.BottomNavTab
-import com.afternote.feature.afternote.domain.model.ListItem
+import com.afternote.feature.afternote.domain.model.author.ListItem
 import com.afternote.feature.afternote.presentation.shared.AfternoteCategory
 
 /**
